requirements.txt
alabaster==0.7.10
amqp==2.3.2
annoy==1.12.0
appdirs==1.4.3
APScheduler==3.5.1
apt-xapian-index==0.47
apturl==0.5.2
asn1crypto==0.24.0
astroid==1.6.3
awscli==1.15.4
Babel==2.5.3
backcall==0.1.0
beautifulsoup4==4.6.1
billiard==3.5.0.3
bitarray==0.8.2
bleach==2.1.3
blinker==1.3
boto==2.48.0
boto3==1.7.19
botocore==1.10.19
Brlapi==0.6.6
bz2file==0.98
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
distro-info==0.18
docutils==0.14
elasticsearch==6.3.0
-e git+https://github.com/mobolic/facebook-sdk.git@3348fb9d02fc2fdfebf6e5b45f1aad06f532730f#egg=facebook_sdk
Flask==1.0.2
Flask-APScheduler==1.8.0
Flask-WTF==0.14.2
future==0.16.0
gensim==3.4.0
googletrans==2.3.0
h5py==2.8.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.7
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
langdetect==1.0.7
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.2
mysql-connector==2.1.6
NearPy==1.0.0
newsapi-python==0.2.2
nltk==3.3
numpy==1.14.5
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
polyglot==16.7.4
profanity==1.1
profanityfilter==2.0.3
protobuf==3.0.0
pyasn1==0.4.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.1
python-crontab==2.3.4
python-dateutil==2.7.3
python-debian==0.1.32
pytz==2018.5
pywikibot==3.0.dev0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.19.1
requests-unixsocket==0.1.5
rsa==3.4.2
s3transfer==0.1.13
schedule==0.5.0
scikit-learn==0.19.1
scipy==1.1.0
SecretStorage==2.3.1
Shapely==1.6.4.post1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
smart-open==1.5.7
sqlitedict==1.5.0
system-service==0.3
systemd-python==234
tinydb==3.10.0
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.23
usb-creator==0.3.3
vecpy==0.1.6
vine==1.1.4
virtualenv==15.1.0
wadllib==1.3.2
webencodings==0.5.1
Werkzeug==0.14.1
wikipedia==1.4.0
wrapt==1.10.11
WTForms==2.1
xkit==0.0.0
zope.interface==4.3.2
